feat(lib): add support to NextJS 13
