+++ package.json
@@ -8,6 +8,9 @@
     "prebuild": "cp -r ./README.md ./packages/headless-stepper",
     "build:lib-fast": "npm run prebuild && pnpm exec tsup --config packages/headless-stepper/tsup.config.ts",
     "build:lib": "npm run prebuild && pnpm exec tsup --config packages/headless-stepper/tsup.config.ts --dts-resolve",
+    "build:components": "pnpm exec tsup --config packages/headless-stepper/components/tsup.config.ts --dts-resolve",
+    "build:components-fast": "pnpm exec tsup --config packages/headless-stepper/components/tsup.config.ts",
+    "build:all": "pnpm run build:lib && pnpm run build:components",
     "nx": "nx",
     "publish:prod": "npm run build:lib && cd ./packages/headless-stepper && npm publish",
     "publish:dev": "npm run build:lib && cd ./packages/headless-stepper && npm publish --tag next",
