+++ packages/headless-stepper/components/src/components/index.ts
@@ -1,0 +1,2 @@
+export * from './Stepper';
+export * from './Step';
