+++ packages/headless-stepper/components/tsup.config.ts
@@ -1,15 +1,16 @@
 import { defineConfig } from 'tsup';
 import path from 'path';
 
-const p = path.join(__dirname, 'src');
+const entry = path.join(__dirname, 'index.ts');
 
 export default defineConfig({
-  name: 'headless-stepper',
-  entry: [`${p}/index.ts`],
+  name: 'headless-stepper-components',
+  entry: [entry],
   outDir: path.join(__dirname, 'dist'),
   sourcemap: true,
   dts: true,
   format: ['esm', 'cjs'],
   external: ['react'],
   clean: true,
+  onSuccess: path.join(__dirname, 'scripts', 'post-build.sh'),
 });
