+++ packages/headless-stepper/project.json
@@ -3,20 +3,30 @@
   "$schema": "../../node_modules/nx/schemas/project-schema.json",
   "sourceRoot": "packages/headless-stepper/src",
   "projectType": "library",
-  "tags": ["library"],
+  "tags": [
+    "library"
+  ],
   "targets": {
     "build": {
       "executor": "@nrwl/rollup:rollup",
-      "outputs": ["{options.outputPath}"],
+      "outputs": [
+        "{options.outputPath}"
+      ],
       "options": {
         "outputPath": "dist/packages/headless-stepper",
         "tsConfig": "packages/headless-stepper/tsconfig.lib.json",
         "project": "packages/headless-stepper/package.json",
         "entryFile": "packages/headless-stepper/src/index.ts",
-        "external": ["react/jsx-runtime"],
+        "external": [
+          "react",
+          "react/jsx-runtime"
+        ],
         "rollupConfig": "@nrwl/react/plugins/bundle-rollup",
         "compiler": "babel",
-        "format": ["esm", "cjs"],
+        "format": [
+          "esm",
+          "cjs"
+        ],
         "generateExportsField": true,
         "assets": [
           {
@@ -39,14 +49,20 @@
     },
     "lint": {
       "executor": "@nrwl/linter:eslint",
-      "outputs": ["{options.outputFile}"],
+      "outputs": [
+        "{options.outputFile}"
+      ],
       "options": {
-        "lintFilePatterns": ["packages/headless-stepper/**/*.{ts,tsx,js,jsx}"]
+        "lintFilePatterns": [
+          "packages/headless-stepper/**/*.{ts,tsx,js,jsx}"
+        ]
       }
     },
     "test": {
       "executor": "@nrwl/jest:jest",
-      "outputs": ["{workspaceRoot}/coverage/packages/headless-stepper"],
+      "outputs": [
+        "{workspaceRoot}/coverage/packages/headless-stepper"
+      ],
       "options": {
         "jestConfig": "packages/headless-stepper/jest.config.ts",
         "passWithNoTests": true
@@ -67,7 +83,9 @@
     },
     "build-storybook": {
       "executor": "@nrwl/storybook:build",
-      "outputs": ["{options.outputDir}"],
+      "outputs": [
+        "{options.outputDir}"
+      ],
       "options": {
         "uiFramework": "@storybook/react",
         "configDir": "packages/headless-stepper/.storybook",
