+++ packages/headless-stepper/src/index.ts
@@ -1,4 +1,4 @@
 // Public API for stepper.
 export * from './lib/hooks/useStepper';
-
-export * from './lib/components/Stepper';
+export * from './lib/types';
+export { useIsomorphicId } from './lib/hooks/useId';
